+++ test/tabular_approximator.jl
@@ -11,32 +11,32 @@ using Test
     n_action = 10
     optimizer_ = Descent(0.125)
 
-    table_q = zeros(Float32, n_action, n_state)
+    table_q = zeros(Float64, n_action, n_state)
     tabular_q_approx = TabularApproximator(table_q, Descent(0.125))
     @test tabular_q_approx.table == table_q
     @test tabular_q_approx.optimizer isa Descent
     @test tabular_q_approx(10) isa SubArray
-    @test tabular_q_approx(10, 1) isa Float32
+    @test tabular_q_approx(10, 1) isa Float64
     RLBase.optimise!(tabular_q_approx, (10, 1) => 1.0)
     @test tabular_q_approx.table[1, 10] == -0.125f0
     RLBase.optimise!(tabular_q_approx, 11 => fill(1.0, n_action))
     @test tabular_q_approx.table[1, 11] == -0.125f0
 
-    table_v = zeros(Float32, n_state)
+    table_v = zeros(Float64, n_state)
     tabular_v_approx = TabularApproximator(table_v, Descent(0.125))
     @test tabular_v_approx.table == table_v
     @test tabular_v_approx.optimizer isa Descent
-    @test tabular_v_approx(10) isa Float32
+    @test tabular_v_approx(10) isa Float64
     RLBase.optimise!(tabular_v_approx, 10 => 1.0)
     @test tabular_v_approx.table[10] == -0.125f0
 
-    @test TabularVApproximator(; n_state = n_state, init = Float32(0.0)).table ==
-          zeros(Float32, n_state)
+    @test TabularVApproximator(; n_state = n_state, init = Float64(0.0)).table ==
+          zeros(Float64, n_state)
     @test TabularQApproximator(;
         n_state = n_state,
         n_action = n_action,
-        init = Float32(0.0),
-    ).table == zeros(Float32, n_action, n_state)
+        init = Float64(0.0),
+    ).table == zeros(Float64, n_action, n_state)
 end
 
 
@@ -45,16 +45,16 @@ end
     td_learner = TDLearner(;
         # TabularQApproximator with specified init matrix
         approximator = TabularApproximator(
-            zeros(Float32, length(RLBase.action_space(env)), length(state_space(env))),
+            zeros(Float64, length(RLBase.action_space(env)), length(state_space(env))),
             Descent(0.125),
         ),
         method = :SARS,
         γ = 0.95,
         n = 0,
     )
-    @test td_learner(env) == zeros(Float32, RLBase.action_space(env))
-    @test td_learner(10) == zeros(Float32, RLBase.action_space(env))
-    @test td_learner(10, 1) == Float32(0)
+    @test td_learner(env) == zeros(Float64, RLBase.action_space(env))
+    @test td_learner(10) == zeros(Float64, RLBase.action_space(env))
+    @test td_learner(10, 1) == Float64(0)
 
     nt_ = (; state = [fill(2, 2)], action = [fill(2, 2)], reward = 0.31, terminal = true)
 
@@ -77,7 +77,7 @@ end
             # TabularQApproximator with specified init matrix
             approximator = TabularApproximator(
                 zeros(
-                    Float32,
+                    Float64,
                     length(RLBase.action_space(env)),
                     length(RLBase.state_space(env)),
                 ),
@@ -105,7 +105,7 @@ end
                     # TabularQApproximator with specified init matrix
                     approximator = TabularApproximator(
                         zeros(
-                            Float32,
+                            Float64,
                             length(RLBase.action_space(env)),
                             length(RLBase.state_space(env)),
                         ),
@@ -122,7 +122,7 @@ end
                     capacity = 3,
                     state = Int64 => (225,),
                     action = Int64 => (15,),
-                    reward = Float32 => (),
+                    reward = Float64 => (),
                     terminal = Bool => (),
                 ),
                 DummySampler(),
@@ -134,7 +134,7 @@ end
                     # TabularQApproximator with specified init matrix
                     approximator = TabularApproximator(
                         zeros(
-                            Float32,
+                            Float64,
                             length(RLBase.action_space(env)),
                             length(RLBase.state_space(env)),
                         ),
@@ -151,7 +151,7 @@ end
                     capacity = 3,
                     state = Int64 => (225,),
                     action = Int64 => (15,),
-                    reward = Float32 => (),
+                    reward = Float64 => (),
                     terminal = Bool => (),
                 ),
                 DummySampler(),
